Split OPL emulation into two sets of files
